feat: updatet services with transaction function

+++ src/currency_exchange/api/ischemas.py
@@ -1,9 +1,12 @@
+from decimal import Decimal
 from typing import Optional
 from pydantic import BaseModel
 
 
 class ExchangePairArgs:
-    def __init__(self, exchange_from: str, exchange_to: str, amount: int | None = None):
+    def __init__(
+        self, exchange_from: str, exchange_to: str, amount: Decimal | None = None
+    ):
         self.exchange_from = exchange_from
         self.exchange_to = exchange_to
         self.amount = amount if amount else 1
